+++ LittleHampton/MHSystems/app/src/main/java/com/mh/systems/littlehampton/ui/activites/DashboardActivity.java
@@ -261,7 +261,7 @@ public class DashboardActivity extends BaseActivity {
             dashboardItemsArrayList.add(new DashboardItems(
                     R.mipmap.ic_home_diary,
                     "Course Diary",
-                    getApplicationContext().getPackageName() + ".ui.activites.CourseDiaryActivity"));
+                    getApplicationContext().getPackageName() + ".ui.activites.CourseDiaryWebviewActivity"));
         }
 
         //Add Competitions