+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/ui/activites/BaseActivity.java
@@ -423,4 +423,18 @@ public class BaseActivity extends AppCompatActivity {
         //If permission is not granted returning false
         return false;
     }
+
+    /**
+     * To report exception on Roll-bar with user
+     * information.
+     *
+     * @param strNameOfScreen : Name of the Activity/Fragment.
+     * @param strException    : Cause of Exception.
+     */
+    public void reportRollBarException(String strNameOfScreen, String strException) {
+        Rollbar.setPersonData(loadPreferenceValue(ApplicationGlobal.KEY_MEMBERID, "N/A"),
+                loadPreferenceValue(ApplicationGlobal.KEY_USER_LOGINID, "N/A"),
+                null);
+        Rollbar.reportMessage(strNameOfScreen, strException);
+    }
 }