+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/ui/activites/DashboardActivity.java
@@ -1,9 +1,12 @@
 package com.mh.systems.demoapp.ui.activites;
 
+import android.app.AlertDialog;
 import android.content.BroadcastReceiver;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
@@ -25,6 +28,7 @@ import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
 import com.mh.systems.demoapp.R;
 import com.mh.systems.demoapp.ui.adapter.RecyclerAdapter.DashboardRecyclerAdapter;
+import com.mh.systems.demoapp.utils.SyncMarket;
 import com.mh.systems.demoapp.utils.constants.ApplicationGlobal;
 import com.mh.systems.demoapp.web.models.deletetoken.AJsonParamsDeleteToken;
 import com.mh.systems.demoapp.web.models.deletetoken.DeleteTokenAPI;
@@ -140,7 +144,7 @@ public class DashboardActivity extends BaseActivity {
      *******************************/
     ArrayList<DashboardItems> dashboardItemsArrayList = new ArrayList<>();
 
-    int iHandicapPosition = -1;
+    int iHandicapPosition = 0;
     String strNameOfWeatherLoc = "";
 
     @Override
@@ -153,6 +157,12 @@ public class DashboardActivity extends BaseActivity {
          * automatically cast the corresponding view in your layout.
          */
         ButterKnife.bind(DashboardActivity.this);
+
+        try {
+            checkUpdateVersion();
+        } catch (PackageManager.NameNotFoundException e) {
+            e.printStackTrace();
+        }
 
         //Initialize adapter.
         dashboardRecyclerAdapter = new DashboardRecyclerAdapter(this,
@@ -286,6 +296,50 @@ public class DashboardActivity extends BaseActivity {
          */
         public void setStrTitleOfGrid(String strTitleOfGrid) {
             this.strTitleOfGrid = strTitleOfGrid;
+        }
+    }
+
+    /**
+     * Check if any update version available on
+     * Google Store.
+     */
+    private void checkUpdateVersion() throws PackageManager.NameNotFoundException {
+        SyncMarket.Initialize(this);
+        final String version = getPackageManager().getPackageInfo(getPackageName(), 0).versionName;
+
+        if(SyncMarket.getMarketVersion() != null && SyncMarket.getMarketVersion().equals(version)){
+            savePreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, version);
+        }
+
+        if (loadPreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, "").equals("")) {
+            savePreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, version);
+        } else {
+            if (!SyncMarket.isVersionEqual(loadPreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, "1.0"))
+                    && SyncMarket.getMarketVersion() != null) {
+                AlertDialog.Builder alertDlg = new AlertDialog.Builder(this);
+                alertDlg.setTitle("Alert");
+                alertDlg.setPositiveButton("Update now", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.dismiss();
+                        SyncMarket.gotoMarket();
+                    }
+                });
+
+                alertDlg.setNegativeButton("Not now", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        //Store latest version to {@link SharedPreferences} so that never alert for current update.
+                        savePreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, SyncMarket.getMarketVersion());
+                        dialog.dismiss();
+                    }
+                });
+
+                alertDlg.setMessage(String.format("A New version of " +
+                        getString(R.string.title_golf_club_name) + " (" + SyncMarket.getMarketVersion()
+                        + ") is now available. Please press “UPDATE NOW” to update your current version"));
+                alertDlg.show();
+            }
         }
     }