+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/ui/activites/MemberDetailActivity.java
@@ -18,6 +18,7 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -68,6 +69,7 @@ public class MemberDetailActivity extends BaseActivity {
 
     String strAddressLine, strMemberEmail, strTelNoHome, strTelNoWork, strTelNoMob, strHandCapPlay;
     String strNameOfMember;
+    String strMemberShipType;
     int iMemberID;
 
     /**
@@ -88,6 +90,7 @@ public class MemberDetailActivity extends BaseActivity {
      *******************************/
     FloatingActionButton fabFriendInvitation;
     FrameLayout flEmailGroup, flContactGroup, flAddressGroup, flWorkGroup, flHomeGroup;
+    LinearLayout llMemberShipType;
     TextView tvMemberNameDD, tvMobContact, tvWorkContact, tvHomeContact, tvMemberEmail, tvMemberAddress, tvMemberJoinDate, tvHandicapPlayStr, tvHandicapTypeStr;
     ImageView ivActionMap, ivActionEmail, ivActionCall;
     Toolbar tbMemberDetail;
@@ -286,6 +289,8 @@ public class MemberDetailActivity extends BaseActivity {
         flWorkGroup = (FrameLayout) findViewById(R.id.flWorkGroup);
         flHomeGroup = (FrameLayout) findViewById(R.id.flHomeGroup);
         flAddressGroup = (FrameLayout) findViewById(R.id.flAddressGroup);
+
+        llMemberShipType = (LinearLayout) findViewById(R.id.llMemberShipType);
 
         fabFriendInvitation = (FloatingActionButton) findViewById(R.id.fabFriendInvitation);
 
@@ -508,6 +513,8 @@ public class MemberDetailActivity extends BaseActivity {
 
                     updateIsFriendUI(membersDetailItems.getData().getIsfriend());
 
+                    strMemberShipType = membersDetailItems.getData().getMembershipStatus();
+
                     displayMembersData();
                 } else {
                     showAlertMessage(getResources().getString(R.string.error_no_data));
@@ -595,7 +602,16 @@ public class MemberDetailActivity extends BaseActivity {
             tvHandicapPlayStr.setText(strHandCapPlay);
         }
 
-        tvHandicapTypeStr.setText(membersDetailItems.getData().getMembershipStatus());
+        /**
+         * Display Membership Type if club allow.
+         */
+        if (strMemberShipType.length() > 0) {
+            llMemberShipType.setVisibility(View.VISIBLE);
+            tvHandicapTypeStr.setText(strMemberShipType);
+        } else {
+            llMemberShipType.setVisibility(View.GONE);
+        }
+
         tvMemberNameDD.setText(strNameOfMember);
         tvMemberJoinDate.setText(getResources().getString(R.string.text_member_since) + " " + getFormateDate(membersDetailItems.getData().getStrLastJoiningDate()));