+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/web/models/TransactionListData.java
@@ -20,12 +20,19 @@ public class TransactionListData {
     @SerializedName("AmountStr")
     @Expose
     private String AmountStr;
+    @SerializedName("BalanceStr")
+    @Expose
+    private String BalanceStr;
     @SerializedName("DiscountTitle")
     @Expose
     private String DiscountTitle;
     @SerializedName("DiscountAmountStr")
     @Expose
     private String DiscountAmountStr;
+    @SerializedName("DiscountBalanceStr")
+    @Expose
+
+    private String DiscountBalanceStr;
     @SerializedName("DateStr")
     @Expose
     private String DateStr;
@@ -37,129 +44,200 @@ public class TransactionListData {
     private Integer TransactionId;
 
     /**
-     * @return The CardStr
+     * 
+     * @return
+     *     The CardStr
      */
     public String getCardStr() {
         return CardStr;
     }
 
     /**
-     * @param CardStr The CardStr
+     * 
+     * @param CardStr
+     *     The CardStr
      */
     public void setCardStr(String CardStr) {
         this.CardStr = CardStr;
     }
 
     /**
-     * @return The TimeStr
+     * 
+     * @return
+     *     The TimeStr
      */
     public String getTimeStr() {
         return TimeStr;
     }
 
     /**
-     * @param TimeStr The TimeStr
+     * 
+     * @param TimeStr
+     *     The TimeStr
      */
     public void setTimeStr(String TimeStr) {
         this.TimeStr = TimeStr;
     }
 
     /**
-     * @return The Title
+     * 
+     * @return
+     *     The Title
      */
     public String getTitle() {
         return Title;
     }
 
     /**
-     * @param Title The Title
+     * 
+     * @param Title
+     *     The Title
      */
     public void setTitle(String Title) {
         this.Title = Title;
     }
 
     /**
-     * @return The AmountStr
+     * 
+     * @return
+     *     The AmountStr
      */
     public String getAmountStr() {
         return AmountStr;
     }
 
     /**
-     * @param AmountStr The AmountStr
+     * 
+     * @param AmountStr
+     *     The AmountStr
      */
     public void setAmountStr(String AmountStr) {
         this.AmountStr = AmountStr;
     }
 
     /**
-     * @return The DiscountTitle
+     * 
+     * @return
+     *     The DiscountTitle
      */
     public String getDiscountTitle() {
         return DiscountTitle;
     }
 
     /**
-     * @param DiscountTitle The DiscountTitle
+     * 
+     * @param DiscountTitle
+     *     The DiscountTitle
      */
     public void setDiscountTitle(String DiscountTitle) {
         this.DiscountTitle = DiscountTitle;
     }
 
     /**
-     * @return The DiscountAmountStr
+     * 
+     * @return
+     *     The DiscountAmountStr
      */
     public String getDiscountAmountStr() {
         return DiscountAmountStr;
     }
 
     /**
-     * @param DiscountAmountStr The DiscountAmountStr
+     * 
+     * @param DiscountAmountStr
+     *     The DiscountAmountStr
      */
     public void setDiscountAmountStr(String DiscountAmountStr) {
         this.DiscountAmountStr = DiscountAmountStr;
     }
 
     /**
-     * @return The DateStr
+     * 
+     * @return
+     *     The DateStr
      */
     public String getDateStr() {
         return DateStr;
     }
 
     /**
-     * @param DateStr The DateStr
+     * 
+     * @param DateStr
+     *     The DateStr
      */
     public void setDateStr(String DateStr) {
         this.DateStr = DateStr;
     }
 
     /**
-     * @return The IsTopup
+     * 
+     * @return
+     *     The IsTopup
      */
     public Boolean getIsTopup() {
         return IsTopup;
     }
 
     /**
-     * @param IsTopup The IsTopup
+     * 
+     * @param IsTopup
+     *     The IsTopup
      */
     public void setIsTopup(Boolean IsTopup) {
         this.IsTopup = IsTopup;
     }
 
     /**
-     * @return The TransactionId
+     * 
+     * @return
+     *     The TransactionId
      */
     public Integer getTransactionId() {
         return TransactionId;
     }
 
     /**
-     * @param TransactionId The TransactionId
+     * 
+     * @param TransactionId
+     *     The TransactionId
      */
     public void setTransactionId(Integer TransactionId) {
         this.TransactionId = TransactionId;
     }
 
+    /**
+     *
+     * @return
+     *     The BalanceStr
+     */
+    public String getBalanceStr() {
+        return BalanceStr;
+    }
+
+    /**
+     *
+     * @param BalanceStr
+     *     The BalanceStr
+     */
+    public void setBalanceStr(String balanceStr) {
+        BalanceStr = balanceStr;
+    }
+
+    /**
+     *
+     * @return
+     *     The DiscountBalanceStr
+     */
+    public String getDiscountBalanceStr() {
+        return DiscountBalanceStr;
+    }
+
+    /**
+     *
+     * @param DiscountBalanceStr
+     *     The DiscountBalanceStr
+     */
+    public void setDiscountBalanceStr(String discountBalanceStr) {
+        DiscountBalanceStr = discountBalanceStr;
+    }
 }