+++ MHS Demo/MHSystems/CHANGELOG.md
@@ -10,7 +10,12 @@
  
  ####15th May, 2017
  - FIXED: android.view.WindowManager$BadTokenException: Unable to add window -- token android.os.BinderProxy@5878646 is not valid; is your activity running?
-    
+ 
+ ####14th June, 2017
+   - UPDATE: Update the Finance tab UI with 4 columns (date/time, item, amount and balance).
+   - FIXED : Navigate back from finance detail to finance tab, top-right filter icon was not working. Now fixed.
+   - UPDATE: Add In App Update Notification version alert dialog one time when any update found.
+   - UPDATE: Integrate the feature flag to Membership Type in Members.
  
 ####Version 1.0 (2017-03-17)