+++ Woolston Manor/MHSystems/CHANGELOG.md
@@ -1,3 +1,3 @@
-##Version 1.0.0 ()
+##Version 1.0.0 (14th June, 2017)
 
 Initial release.